Updated Bottom menu, adapters

+++ app/src/main/kotlin/com/popalay/yooder/managers/FirebaseManager.kt
@@ -59,5 +59,6 @@ class FirebaseManager(val ref: Firebase) : DataManager {
     }
 
     override fun getMyRemindersQuery(myId: String) = ref.child("reminders").orderByChild("from").equalTo(myId)
-    override fun getMyNotificationsQuery(myId: String) = ref.child("reminders").orderByChild("to").equalTo(myId)
+
+    override fun getMyNotificationsQuery(myId: String) = ref.child("reminders").orderByChild("to").equalTo(myId.toDouble())
 }
